ApiUtils - remove platform specific implementations and use a single file
